feat: refact using GlobalSections
