+++ packages/core/src/Layout.tsx
@@ -12,7 +12,30 @@ const CartSidebar = lazy(() => import('src/components/cart/CartSidebar'))
 const RegionModal = lazy(() => import('src/components/region/RegionModal'))
 
 function Layout({ children }: PropsWithChildren) {
-  return <>{children}</>
+  const { cart: displayCart, modal: displayModal } = useUI()
+
+  return (
+    <>
+      <main>
+        <RegionBar className="display-mobile" />
+        {children}
+      </main>
+
+      <Footer />
+
+      {displayCart && (
+        <Suspense fallback={null}>
+          <CartSidebar />
+        </Suspense>
+      )}
+
+      {displayModal && (
+        <Suspense fallback={null}>
+          <RegionModal />
+        </Suspense>
+      )}
+    </>
+  )
 }
 
 export default Layout
